requirements.txt
tensorflow-macos
tensorflow-metal
pandas
scikit-learn
jupyterlab
opencv-python-headless
Pillow